+++ src/engine/Actions/RotationType.ts
@@ -6,20 +6,20 @@ export enum RotationType {
    * Rotation via `ShortestPath` will use the smallest angle
    * between the starting and ending points. This strategy is the default behavior.
    */
-  ShortestPath = 0,
+  ShortestPath = 'shortest-path',
   /**
    * Rotation via `LongestPath` will use the largest angle
    * between the starting and ending points.
    */
-  LongestPath = 1,
+  LongestPath = 'longest-path',
   /**
    * Rotation via `Clockwise` will travel in a clockwise direction,
    * regardless of the starting and ending points.
    */
-  Clockwise = 2,
+  Clockwise = 'clockwise',
   /**
    * Rotation via `CounterClockwise` will travel in a counterclockwise direction,
    * regardless of the starting and ending points.
    */
-  CounterClockwise = 3
+  CounterClockwise = 'counter-clockwise'
 }